Added to_sensor API and renamed original file
Outlined the API calls to the sensor itself.

Added an example property of response_topic as a placeholder if integration of two topic approach is non-viable through AWS or Azure

The original payload API was a js file and was renamed to a json file with from_sensor to clarify purpose.

Changed command_id to correlation_id to more clearly outline its primary purpose.

Added comments suggesting migration to string for correlation id.